[refactor] Move FunctionEvaluator to pinot-spi; Groovy/TimeSpec evaluators to pinot-common
- FunctionEvaluator interface promoted to pinot-spi.function (depends only on GenericRow)
- GroovyFunctionEvaluator, GroovyStaticAnalyzerConfig, TimeSpecFunctionEvaluator moved to
  pinot-common.evaluator as canonical implementations
- pinot-segment-spi.function.FunctionEvaluator becomes a @Deprecated forwarding shim
- Groovy/TimeSpec classes removed from pinot-segment-spi entirely (avoids duplicate static
  security state that would silently bypass sandbox configuration)
- pinot-segment-local.function shims updated to extend pinot-common.evaluator classes
- All first-party callers (broker, controller, core, segment-local) updated to new packages
