+++ pinot-common/src/main/java/org/apache/pinot/common/evaluator/GroovyStaticAnalyzerConfig.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.segment.spi.function;
+package org.apache.pinot.common.evaluator;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
