+++ pinot-core/src/main/java/org/apache/pinot/core/segment/processing/partitioner/TransformFunctionPartitioner.java
@@ -19,8 +19,8 @@
 package org.apache.pinot.core.segment.processing.partitioner;
 
 import org.apache.pinot.common.evaluator.FunctionEvaluatorFactory;
-import org.apache.pinot.segment.spi.function.FunctionEvaluator;
 import org.apache.pinot.spi.data.readers.GenericRow;
+import org.apache.pinot.spi.function.FunctionEvaluator;
 
 
 /**
