+++ pinot-core/src/test/java/org/apache/pinot/core/data/function/GroovyFunctionEvaluatorTest.java
@@ -23,15 +23,15 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.apache.pinot.segment.spi.function.GroovyFunctionEvaluator;
-import org.apache.pinot.segment.spi.function.GroovyStaticAnalyzerConfig;
+import org.apache.pinot.common.evaluator.GroovyFunctionEvaluator;
+import org.apache.pinot.common.evaluator.GroovyStaticAnalyzerConfig;
 import org.apache.pinot.spi.data.readers.GenericRow;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import org.testng.collections.Lists;
 
-import static org.apache.pinot.segment.spi.function.GroovyStaticAnalyzerConfig.getDefaultAllowedImports;
-import static org.apache.pinot.segment.spi.function.GroovyStaticAnalyzerConfig.getDefaultAllowedReceivers;
+import static org.apache.pinot.common.evaluator.GroovyStaticAnalyzerConfig.getDefaultAllowedImports;
+import static org.apache.pinot.common.evaluator.GroovyStaticAnalyzerConfig.getDefaultAllowedReceivers;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.fail;
