+++ pinot-core/src/test/java/org/apache/pinot/core/data/function/GroovyStaticAnalyzerConfigTest.java
@@ -20,7 +20,7 @@ package org.apache.pinot.core.data.function;
 
 import java.util.Iterator;
 import java.util.List;
-import org.apache.pinot.segment.spi.function.GroovyStaticAnalyzerConfig;
+import org.apache.pinot.common.evaluator.GroovyStaticAnalyzerConfig;
 import org.apache.pinot.spi.utils.JsonUtils;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
