+++ pinot-segment-spi/src/main/java/org/apache/pinot/segment/spi/function/FunctionEvaluator.java
@@ -18,27 +18,15 @@
  */
 package org.apache.pinot.segment.spi.function;
 
-import java.util.List;
-import org.apache.pinot.spi.data.readers.GenericRow;
-
 
 /**
- * Interface for evaluators of transform function expressions of schema field specs
+ * Deprecated forwarding interface that preserves the historical
+ * {@code org.apache.pinot.segment.spi.function} binary name for downstream callers.
+ *
+ * <p>This interface is stateless and thread-safe.
+ *
+ * @deprecated Use {@link org.apache.pinot.spi.function.FunctionEvaluator} instead.
  */
-public interface FunctionEvaluator {
-
-  /**
-   * Get the arguments of the function
-   */
-  List<String> getArguments();
-
-  /**
-   * Evaluate the function on the generic row and return the result
-   */
-  Object evaluate(GenericRow genericRow);
-
-  /**
-   * Evaluates the function on the given values (same order as the arguments) and returns the result.
-   */
-  Object evaluate(Object[] values);
+@Deprecated
+public interface FunctionEvaluator extends org.apache.pinot.spi.function.FunctionEvaluator {
 }
