+++ pinot-udf-test/src/main/java/org/apache/pinot/udf/test/scenarios/ExpressionTransformerTestScenario.java
@@ -25,8 +25,8 @@ import org.apache.pinot.common.evaluator.FunctionEvaluatorFactory;
 import org.apache.pinot.core.udf.Udf;
 import org.apache.pinot.core.udf.UdfExample;
 import org.apache.pinot.core.udf.UdfSignature;
-import org.apache.pinot.segment.spi.function.FunctionEvaluator;
 import org.apache.pinot.spi.data.readers.GenericRow;
+import org.apache.pinot.spi.function.FunctionEvaluator;
 import org.apache.pinot.udf.test.PinotFunctionEnvGenerator;
 import org.apache.pinot.udf.test.UdfExampleResult;
 import org.apache.pinot.udf.test.UdfTestCluster;
